chore: Split e2e tests into fixtures for speed
