+++ test/e2e/move/new/stc/to-court.test.js
@@ -1,0 +1,62 @@
+import { Selector } from 'testcafe'
+
+import { stcUser } from '../../../_roles'
+import { newMove } from '../../../_routes'
+import { page, moveDetailPage, createMovePage } from '../../../pages'
+
+fixture('New move from Secure Training Centre (STC) to Court').beforeEach(
+  async t => {
+    await t.useRole(stcUser).navigateTo(newMove)
+  }
+)
+
+test('With a new person', async t => {
+  // PNC lookup
+  await t
+    .expect(page.getCurrentUrl())
+    .contains('/person-lookup-pnc')
+    .click(createMovePage.steps.personLookup.nodes.noIdentifierLink)
+    .click(createMovePage.steps.personLookup.nodes.moveSomeoneNew)
+
+  // Personal details
+  const personalDetails = await createMovePage.fillInPersonalDetails()
+  await page.submitForm()
+
+  // Move details
+  const moveDetails = await createMovePage.fillInMoveDetails('Court')
+  await page.submitForm()
+
+  // Move date
+  await createMovePage.fillInDate()
+  await page.submitForm()
+
+  // Court information
+  const courtInformation = await createMovePage.fillInCourtInformation()
+  await page.submitForm()
+
+  // Confirmation page
+  await createMovePage.checkConfirmationStep({
+    fullname: personalDetails.fullname,
+    location: moveDetails.courtLocation,
+  })
+  await t.click(Selector('a').withExactText(personalDetails.fullname))
+
+  await t.click(Selector('a').withExactText('Details'))
+
+  // Move detail assertions
+  await moveDetailPage.checkHeader({ fullname: personalDetails.fullname })
+
+  // Personal details assertions
+  await moveDetailPage.checkPersonalDetails(personalDetails)
+
+  // Check assessment
+  await moveDetailPage.checkCourtInformation(courtInformation)
+
+  // Check banner
+  await t
+    .expect(moveDetailPage.nodes.identityBar.innerText)
+    .contains(
+      'Start Youth Risk Assessment',
+      'Should contain start assessment banner'
+    )
+})
