+++ test/e2e/move/timeline.test.js
@@ -1,8 +1,8 @@
-import { expectStatusCode } from './_helpers'
-import { createMove } from './_move'
-import { policeUser } from './_roles'
-import { home, getTimeline } from './_routes'
-import { moveTimelinePage } from './pages'
+import { expectStatusCode } from '../_helpers'
+import { createMove } from '../_move'
+import { policeUser } from '../_roles'
+import { home, getTimeline } from '../_routes'
+import { moveTimelinePage } from '../pages'
 
 fixture('Move timeline').beforeEach(async t => {
   await t.useRole(policeUser).navigateTo(home)
