+++ test/e2e/move/update/forbidden.test.js
@@ -1,0 +1,25 @@
+import { getRandomLocation } from '../../_helpers'
+import {
+  createCourtMove,
+  checkNoUpdateLinks,
+  checkUpdatePagesForbidden,
+} from '../../_move'
+import { supplierUser } from '../../_roles'
+import { home } from '../../_routes'
+
+fixture('Existing move (As supplier user)').beforeEach(async t => {
+  await t.useRole(supplierUser).navigateTo(home)
+  await createCourtMove({
+    moveOverrides: {
+      from_location: await getRandomLocation('police', {
+        shouldHaveSupplier: true,
+      }),
+    },
+  })
+})
+
+test('User should not be able to update a move', async () => {
+  await checkNoUpdateLinks()
+
+  await checkUpdatePagesForbidden()
+})
