+++ test/e2e/move/update/police/to-court.test.js
@@ -1,12 +1,10 @@
 import { addDays, format } from 'date-fns'
 import { Selector } from 'testcafe'
 
-import { acceptMove, startMove } from './_helpers'
 import {
   createCourtMove,
   checkUpdateLinks,
   checkUpdatePagesAccessible,
-  checkUpdatePagesRedirected,
   checkPoliceNationalComputerReadOnly,
   checkUpdatePersonalDetails,
   checkUpdateRiskInformation,
@@ -14,9 +12,9 @@ import {
   checkUpdateCourtInformation,
   // checkUpdateMoveDetails,
   checkUpdateMoveDate,
-} from './_move'
-import { policeUser } from './_roles'
-import { home, getMove } from './_routes'
+} from '../../../_move'
+import { policeUser } from '../../../_roles'
+import { home, getMove } from '../../../_routes'
 
 fixture('Existing move from Police Custody to Court').beforeEach(async t => {
   await t.useRole(policeUser).navigateTo(home)
@@ -69,43 +67,8 @@ test.before(async t => {
   })
   await t.click(Selector('a').withExactText('Details'))
 })('Without existing PNC number', async t => {
+  // TODO: re-enable this test, disabled in https://github.com/ministryofjustice/hmpps-book-secure-move-frontend/commit/edbe42ecfb91139355f319468fa3a5c9ce729ab4
   // await checkUpdatePersonalDetails({
   //   include: ['policeNationalComputer'],
   // })
-})
-
-fixture.beforeEach(async t => {
-  await t.useRole(policeUser).navigateTo(home)
-  await createCourtMove({
-    moveOverrides: {
-      move_type: 'prison_recall',
-    },
-  })
-  await t.click(Selector('a').withExactText('Details'))
-})('Existing move from Police Custody to Prison (recall)')
-
-test('User should be able to update move details', async t => {
-  // await checkUpdateMoveDetails()
-})
-
-fixture.beforeEach(async t => {
-  await t.useRole(policeUser).navigateTo(home)
-  t.ctx.move = await createCourtMove()
-  await acceptMove(t.ctx.move.id)
-  await startMove(t.ctx.move.id)
-  await t.navigateTo(getMove(t.ctx.move.id))
-  await t.click(Selector('a').withExactText('Details'))
-})('Existing move that has left custody')
-
-test('User should not be able to update any information', async t => {
-  // No edit links should be visible
-  await checkUpdateLinks(undefined, true)
-  await checkUpdatePagesRedirected(t.ctx.move.id, [
-    'personal_details',
-    'risk',
-    'health',
-    'court',
-    'move',
-    'date',
-  ])
 })
